Vifm: enhacned filetype and fileviewer options for non-tar gz,bz2,zx archives.
After awesome answer of xaizek (author of VIFM) here:

    https://q2a.vifm.info/2264/look-behind-in-filetype-regex-pattern

pattern was replaced with more vifm-ic one:

    before:
    \/([^r]|[^a]r|[^t]ar|[^\.]tar)\.(gz|bz2|xz)$/

    after:
    \{*.gz}!{*.tar.gz},
    \{*.bz2}!{*.tar.bz2},
    \{*.xz}!{*.tar.xz},

    possible alternative less readable more flexible (may be next time):
    \/\.(gz|bz2|xz)$/!/\.tar\.(gz|bz2|xz)$/

Notes from xaizek's answer:
Several patterns with comma work as disjunction (OR), technically

    filetype {a},{b} command

is just a shorthand for

    filetype {a} command
    filetype {b} command

But patterns without commas work as conjunction (AND), so it is possible
to negate each pattern individually.

See more in :help vifm-patterns.